Move get_code_error_message to new utils file
